Add an extension point to allow the override of the StaticFileServer
The old StaticFileServer has been renamed to DefaultStaticFileServer,
and the StaticFileServer has been converted to an interface with two
methods.
